refactor: extract core fingerprint hashing
